Reorganized so that auth specific utils functions are in the auth blueprint directory and cleaned up use of login_serializer.
