+++ pubs_ui/auth/utils.py
@@ -1,10 +1,11 @@
-
 from urlparse import urlparse, urljoin
 from urllib import unquote
 
 from werkzeug.exceptions import NotFound, MethodNotAllowed
 
-from . import app
+from . import login_serializer
+from .. import app
+
 
 def get_url_endpoint(url, server_name, fallback, wsgi_str=None):
     '''
@@ -39,3 +40,24 @@ def is_safe_url(target, host_url):
     ref_url = urlparse(host_url)
     test_url = urlparse(urljoin(host_url, target))
     return test_url.scheme in ('http', 'https') and ref_url.netloc == test_url.netloc
+
+def generate_auth_header(request):
+    """
+    This is used to generate the auth header to make requests back to the pubs-services endpoints
+    :param request: the request object to get the cookie
+    :return: A authorization header that can be sent along to the pubs-services endpoint
+    """
+    # get the token cookie from the request
+    token_cookie = request.cookies.get('remember_token')
+    # set a max age variable that is the same max age as the cookie can be.
+    max_age = app.config["REMEMBER_COOKIE_DURATION"].total_seconds()
+    # decrypt the cookie to get the username and the token
+    session_data = login_serializer.loads(token_cookie, max_age=max_age)
+    # get the token from the session data
+    mypubs_token = session_data[1]
+    # build the auth value to send to the manage server
+    auth_value = 'Bearer  '+mypubs_token
+    # build the Authorization header
+    header = {'Authorization': auth_value}
+    return header
+
